+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Division by zero when image dimensions not available in video RSS feed
+
 ## [1.3.0] - 2020-12-29
 
 ### Added